Add tables and images to README

+++ mnist_toy/src/nnVector.h
@@ -10,7 +10,7 @@
 
 #include <vector>
 
-using nn_cal_type = float;
+using nn_cal_type = double;
 using vector_1d = std::vector<nn_cal_type>;
 using vector_2d = std::vector< vector_1d >;
 using vector_3d = std::vector< vector_2d >;
